Add ruff/mypy, fix real gaps they found, correct a stale claim about router fallback coverage
Third professional-polish item. ruff (27 findings, mostly Python-version
type-hint modernization, auto-fixed) and mypy (12 findings, all real once
triaged) added with config in pyproject.toml.

mypy found: two Optional-field gaps in the parked V2 Gemini path
(rag.py, index_documents.py) that could genuinely crash on a
safety-filter-blocked or empty API response; a variable-name collision
in local_qa.py that, once renamed to satisfy mypy, exposed a real
NameError waiting to happen on a reachable code path (caught by
re-testing after the edit, not assumed safe); and several llama-cpp-
python stub-typing gaps (its create_chat_completion() types against a
streaming/non-streaming union we never select into) suppressed narrowly
with justified inline comments, not blanket ignores.

While re-verifying the NameError fix end-to-end, found and corrected a
stale claim from 2026-08-07 that "how fast am I going" was already
routing correctly - it wasn't; the router classifies it correctly but
the vehicle-term safety net has no "fast" keyword and blocks the
decision. Attempted fix (adding "fast" as a keyword) caused real false
positives ("fasting", "breakfast") and was reverted - left as an
honestly tracked open gap rather than shipped half-safe.

+++ src/resort_atv_voice/can_telemetry.py
@@ -24,7 +24,7 @@ FIELD_TO_CAN_ID = {field: can_id for can_id, field in CAN_ID_TO_FIELD.items()}
 def encode_frame(field: str, value: float) -> can.Message:
     can_id = FIELD_TO_CAN_ID[field]
     return can.Message(
-        arbitration_id=can_id, data=bytes([int(round(value)) & 0xFF]), is_extended_id=False
+        arbitration_id=can_id, data=bytes([round(value) & 0xFF]), is_extended_id=False
     )
